lms-backend/requirements.txt
Django>=5.0,<6
djangorestframework
django-environ